[ISSUE-3] - added CommandHandler. Moved handlers for commands from CommandService to CommandHandler
