+++ server/CommandHandler/appsettings.Development.json
@@ -1,0 +1,35 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "rabbitmq": {
+    "Username": "admin",
+    "Password": "admin",
+    "VirtualHost": "/",
+    "Port": 5672,
+    "Hostnames": [
+      "rabbitmq"
+    ],
+    "RequestTimeout": "00:00:10",
+    "PublishConfirmTimeout": "00:00:01",
+    "RecoveryInterval": "00:00:10",
+    "PersistentDeliveryMode": true,
+    "AutoCloseConnection": true,
+    "AutomaticRecovery": true,
+    "TopologyRecovery": true,
+    "Exchange": {
+      "Durable": true,
+      "AutoDelete": true,
+      "Type": "Topic"
+    },
+    "Queue": {
+      "AutoDelete": true,
+      "Durable": true,
+      "Exclusive": true
+    }
+  }
+}
