+++ server/server.sln
@@ -16,6 +16,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueryService", "QueryServic
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandService", "CommandService\CommandService.csproj", "{37325116-ED6B-44B5-B47D-1DB2F6C45854}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandHandler", "CommandHandler\CommandHandler.csproj", "{16C1EC9C-6268-4682-9325-F9BF29AECA2F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,6 +48,10 @@ Global
 		{37325116-ED6B-44B5-B47D-1DB2F6C45854}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5E0A68A3-BF24-405B-8226-6337BAA8ECB3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5E0A68A3-BF24-405B-8226-6337BAA8ECB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16C1EC9C-6268-4682-9325-F9BF29AECA2F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16C1EC9C-6268-4682-9325-F9BF29AECA2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16C1EC9C-6268-4682-9325-F9BF29AECA2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16C1EC9C-6268-4682-9325-F9BF29AECA2F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{DE26DC8E-06D1-4FCB-9397-3EE26B9976F6} = {95CB2EAC-5AA2-46BB-81D0-B32EB59E32CE}
